Refactor: 당첨 도메인 추가
