+++ src/main/kotlin/lotto/LottoMain.kt
@@ -1,8 +1,8 @@
 package lotto
 
-import lotto.domain.LottoGame
 import lotto.domain.LottoMachine
 import lotto.domain.RandomLogic
+import lotto.service.LottoGame
 import lotto.view.InputView
 import lotto.view.OutputView
 
@@ -10,15 +10,12 @@ fun main() {
 
     val cash = InputView.inputCash()
     val lottoGame = LottoGame(RandomLogic())
-
     val purchaseLottoList = lottoGame.buyLotto(cash.toInt())
 
     OutputView.showLottoList(purchaseLottoList)
 
     val winningNumberList = InputView.inputWinningNumber()
-
     val winningStatus = lottoGame.getWinningStats(winningNumberList, purchaseLottoList)
-
     val winningRate = LottoMachine.createWinningRate(cash.toInt(), winningStatus)
 
     OutputView.showWinningStatus(winningStatus, winningRate)
